mod_c.rs hands the lowering-context builder and the module splice list to mod_c_tail.rs, a pure text move under the 800-line ceiling
